create some hooks
