+++ src/components/Footer/FooterWithForm.jsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect, useCallback } from "react";
+import { useWindowResize } from "@/hooks/useWindowResize";
 import { socialLinksAndMail } from "@/helpers/linkArrays";
 import CallBtn from "../Buttons/CallBtn/CallBtn";
 import SocialLinksList from "../SocialLinks/SocialLinksList";
@@ -10,57 +10,7 @@ import FooterLinks from "./FooterLinks";
 import styles from "./Footer.module.scss";
 
 const FooterWithForm = () => {
-    const [isMobile, setMobile] = useState(false);
-    const [isLaptop, setLaptop] = useState(false);
-    const [isDesktop, setDesktop] = useState(false);
-
-    const handleResizeMobile = useCallback(() => {
-        if (window.innerWidth < 768) {
-            setMobile(true);
-        } else {
-            setMobile(false);
-        }
-    }, [setMobile]);
-
-    useEffect(() => {
-        window.addEventListener("resize", handleResizeMobile);
-        handleResizeMobile();
-        return () => {
-            window.removeEventListener("resize", handleResizeMobile);
-        };
-    }, [handleResizeMobile]);
-
-    const handleResizeLaptop = useCallback(() => {
-        if (window.innerWidth >= 768 && window.innerWidth < 1024) {
-            setLaptop(true);
-        } else {
-            setLaptop(false);
-        }
-    }, [setLaptop]);
-
-    useEffect(() => {
-        window.addEventListener("resize", handleResizeLaptop);
-        handleResizeLaptop();
-        return () => {
-            window.removeEventListener("resize", handleResizeLaptop);
-        };
-    }, [handleResizeLaptop]);
-
-    const handleResizeDesktop = useCallback(() => {
-        if (window.innerWidth >= 1024) {
-            setDesktop(true);
-        } else {
-            setDesktop(false);
-        }
-    }, [setDesktop]);
-
-    useEffect(() => {
-        window.addEventListener("resize", handleResizeDesktop);
-        handleResizeDesktop();
-        return () => {
-            window.removeEventListener("resize", handleResizeDesktop);
-        };
-    }, [handleResizeDesktop]);
+    const { isMobile, isTablet, isLaptop, isDesktop } = useWindowResize();
 
     return (
         <>
@@ -78,7 +28,7 @@ const FooterWithForm = () => {
                             </div>
                         </footer>
                     );
-                } else if (isLaptop) {
+                } else if (isTablet) {
                     return (
                         <footer className={styles.footerLaptop}>
                             <div className={`container ${styles.container}`}>
@@ -94,7 +44,7 @@ const FooterWithForm = () => {
                             </div>
                         </footer>
                     );
-                } else if (isDesktop) {
+                } else if (isLaptop || isDesktop) {
                     return (
                         <footer className={styles.footerDesktop}>
                             <div className={`container ${styles.container}`}>
